+++ @navikt/core/react/src/list/index.ts
@@ -1,2 +1,2 @@
 export { default as List } from "./List";
-export type { ListProps } from "./List";
+export type { ListProps } from "./types";
